requirements.txt
# dependencies
beautifulsoup4
colorama
duckduckgo_search
md2pdf
playwright
openai
python-dotenv
pyyaml
uvicorn
pydantic
fastapi
python-multipart
markdown
langchain
langchain-openai
langchain-google-genai
langchain_community
tavily-python
arxiv
PyMuPDF
requests
jinja2
aiofiles
newspaper3k
SQLAlchemy
mistune
python-docx
htmldocx
lxml[html_clean]
websockets
unstructured
pytest
pytest-asyncio
langchain-community
langchain-experimental
chromadb
langchainhub
langsmith
PyPDF2